# chacha20: Add NEON implementation for aarch64