--- a/.github/workflows/chacha20.yml
+++ b/.github/workflows/chacha20.yml
@@ -166,7 +166,7 @@ jobs:
           # ARM64 with NEON backend
           - target: aarch64-unknown-linux-gnu
             rust: nightly
-            features: --features nightly
+            features: --features neon
 
           # PPC32
           - target: powerpc-unknown-linux-gnu
@@ -189,5 +189,4 @@ jobs:
       - run: cross test --target ${{ matrix.target }} --release --features force-soft
       - run: cross test --target ${{ matrix.target }} --release --features rng
       - run: cross test --target ${{ matrix.target }} --release --features std
-      # --all-features except nightly
-      - run: cross test --target ${{ matrix.target }} --release --features expose-core,force-soft,hchacha,legacy,rng,std
+      - run: cross test --target ${{ matrix.target }} --release --all-features