--- a/chacha20/Cargo.toml
+++ b/chacha20/Cargo.toml
@@ -37,6 +37,7 @@ expose-core = []
 force-soft = []
 hchacha = ["cipher"]
 legacy = ["cipher"]
+nightly = []
 rng = ["rand_core"]
 std = ["cipher/std"]