Processes four blocks in parallel. Adapted from the SUPERCOP `dolbeau`
backend (public domain).

Placed behind a new `nightly` feature flag, as aarch64 SIMD intrinsics
are not yet stable.