@str4d 2021-08-30
Benchmark environment:
```
$ lscpu
Architecture:        aarch64
Byte Order:          Little Endian
CPU(s):              6
On-line CPU(s) list: 0-5
Thread(s) per core:  1
Core(s) per socket:  3
Socket(s):           2
Vendor ID:           ARM
Model:               4
Model name:          Cortex-A53
Stepping:            r0p4
CPU max MHz:         1896.0000
CPU min MHz:         100.0000
BogoMIPS:            48.00
Flags:               fp asimd evtstrm aes pmull sha1 sha2 crc32
$ cargo +nightly --version
cargo 1.56.0-nightly (f559c109c 2021-08-26)
```

Current master:
```
$ cargo +nightly bench -p chacha20
     Running unittests (target/release/deps/chacha12-206bfb5dd069506d)
test bench1_10     ... bench:          57 ns/iter (+/- 0) = 175 MB/s
test bench2_100    ... bench:         465 ns/iter (+/- 2) = 215 MB/s
test bench3_1000   ... bench:       5,773 ns/iter (+/- 16) = 173 MB/s
test bench4_10000  ... bench:      58,764 ns/iter (+/- 179) = 170 MB/s
test bench5_100000 ... bench:     588,846 ns/iter (+/- 2,404) = 169 MB/s

     Running unittests (target/release/deps/chacha20-8b7c91f66bd44380)
test bench1_10     ... bench:          74 ns/iter (+/- 0) = 135 MB/s
test bench2_100    ... bench:         634 ns/iter (+/- 2) = 157 MB/s
test bench3_1000   ... bench:       7,472 ns/iter (+/- 6) = 133 MB/s
test bench4_10000  ... bench:      75,752 ns/iter (+/- 91) = 132 MB/s
test bench5_100000 ... bench:     758,864 ns/iter (+/- 888) = 131 MB/s

     Running unittests (target/release/deps/chacha8-f812fddd0a2f7553)
test bench1_10     ... bench:          50 ns/iter (+/- 0) = 200 MB/s
test bench2_100    ... bench:         380 ns/iter (+/- 1) = 263 MB/s
test bench3_1000   ... bench:       4,918 ns/iter (+/- 4) = 203 MB/s
test bench4_10000  ... bench:      50,220 ns/iter (+/- 88) = 199 MB/s
test bench5_100000 ... bench:     503,601 ns/iter (+/- 1,495) = 198 MB/s
```

This PR:
```
$ cargo +nightly bench --features nightly -p chacha20
     Running unittests (target/release/deps/chacha12-92e41f0468c669c0)
test bench1_10     ... bench:          56 ns/iter (+/- 1) = 178 MB/s
test bench2_100    ... bench:         280 ns/iter (+/- 0) = 357 MB/s
test bench3_1000   ... bench:       2,298 ns/iter (+/- 5) = 435 MB/s
test bench4_10000  ... bench:      21,815 ns/iter (+/- 204) = 458 MB/s
test bench5_100000 ... bench:     220,187 ns/iter (+/- 806) = 454 MB/s

     Running unittests (target/release/deps/chacha20-7fc4cc5d1c2fd949)
test bench1_10     ... bench:          68 ns/iter (+/- 0) = 147 MB/s
test bench2_100    ... bench:         399 ns/iter (+/- 2) = 250 MB/s
test bench3_1000   ... bench:       3,499 ns/iter (+/- 14) = 285 MB/s
test bench4_10000  ... bench:      33,893 ns/iter (+/- 223) = 295 MB/s
test bench5_100000 ... bench:     339,338 ns/iter (+/- 2,569) = 294 MB/s

     Running unittests (target/release/deps/chacha8-218e8a820d75dd02)
test bench1_10     ... bench:          51 ns/iter (+/- 1) = 196 MB/s
test bench2_100    ... bench:         221 ns/iter (+/- 1) = 452 MB/s
test bench3_1000   ... bench:       1,700 ns/iter (+/- 5) = 588 MB/s
test bench4_10000  ... bench:      15,921 ns/iter (+/- 40) = 628 MB/s
test bench5_100000 ... bench:     160,527 ns/iter (+/- 325) = 622 MB/s
```