@str4d 2021-08-30
This backend should also work for `target_arch = "arm"`, but that requires support in `cpufeatures` for checking NEON support at runtime, and then a fallback to a 4-block wrapper around the soft impl.